fix: num.clamp() returns num, and the analyser was right all along
THE ACTUAL ERROR, finally visible once the LayoutBuilder was out of the way:
  The argument type 'num' can't be assigned to the parameter type 'double?'
  size_grid.dart:112:31

(78 * aspect).clamp(20.0, 62.0) returns num — clamp is defined on num and does
not narrow. Dart will not implicitly convert num to double, so Container(width:)
rejected it.

So the analyser was correct from the very first run. It was reporting a genuine
type error, and my rewrites kept pushing the report to a different call site
until I replaced enough code that the real error could surface.

THREE WRONG THEORIES, all reported as findings:
  1. 'cascade from the unresolved names'  — it survived those fixes
  2. 'Expanded with nothing to expand'    — it survived that fix
  3. 'removing LayoutBuilder fixed it'    — it only moved the error into view

WHAT I SHOULD HAVE DONE FIRST
Checked the types of every expression I was passing to an argument, instead of
reading the shape of the code.  returning  is a two-second fact; I
spent four CI cycles not checking it.

THE RULE THIS EARNS, and it goes in the README:
When a Dart error names a call you believe is correct, do not assume the
analyser is confused — Dart's analyser does not invent phantom errors. Assume
the error is real and the LOCATION is approximate, then check the TYPES of the
expressions at that location before checking the structure.

+++ lib/features/quote/size_grid.dart
@@ -108,8 +108,12 @@ class _Swatch extends StatelessWidget {
                 height: 78,
                 child: Center(
                   child: Container(
-                      width: (78 * aspect).clamp(20.0, 62.0),
-                      height: (78).clamp(20.0, 78.0),
+                      // .toDouble() is not decoration: num.clamp() returns num,
+                      // and Dart will not narrow it to double implicitly. An
+                      // explicit conversion is the honest way to say 'this is a
+                      // double', rather than a cast that could hide a real bug.
+                      width: (78 * aspect).clamp(20.0, 62.0).toDouble(),
+                      height: (78.0).clamp(20.0, 78.0).toDouble(),
                       decoration: BoxDecoration(
                         color: PaperPalette.sheet,
                         border: Border.all(
